Restructured SectionProperties, Added various Concrete Beam types
